consume all events in every iteration of the main loop, clean on exit
The first is necessary (at least in Ubuntu) for the user to be able to close
the main window.  The only way to end the demo is (again, in Ubuntu) to ^C out
of the console where you started it.  That kills the echo in the console (you
can still type commands and they'll work, but you can't see what you're
typing).  Needless to say, that's quite scary.

I don't know whether cleaning up is really necessary in this app, but it can't
hurt to show how to do it.